apps: gameoflife: Added to P8/PineTime builds
I believe that shipping Daniel Thompson's implementation of Conway's Game of Life by default for the time being, instead of providing the user with a simplistic version of the firmware, will help with illustrating the capabilities of wasp-os in a way that's clear and tangible. There's already a video that does illustrate how the Game of Life looks on the PineTime watch (see: https://www.youtube.com/watch?v=YktiGUSRJB4), but I am confident that a filmed demonstration is inferior compared to a hands-on example. I strongly believe that the inclusion of this small game could very well promote adoption, as well as, to an extent, further contributions, without the end user having to go through the alienating hassle of understanding how the codebase works in order to try it out by themself.