README: Move wasp-os community chat to libera.chat
We also update the recommendation for those new to IRC to use
a matrix/IRC bridge instead of a simple IRC web client since that
should make the community a bit more interactive.

If you are curious about the move to libera.chat take a look at
the Linux Weekly News summary of the beginning of the end
(https://lwn.net/Articles/857140/ ) together with these
testimonials about the behaviour of the new freenode staff
which I think makes clear we have reached the end of the end:

 * https://nedbatchelder.com/blog/202106/goodbye_freenode.html
 * https://www.fsf.org/news/update-to-the-fsf-and-gnus-plan-to-move-irc-channels-to-libera.chat
 * https://twitter.com/irccloud/status/1404153550159159298

So... we're moving... and sorry for not making this decision sooner!

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>