Implement hardware-based raise-to-wake gesture with software fallback
With the BMA425 accelerometer shipped with newer PineTime units, it is
possible to configure gesture detection on the hardware and send
interrupts to the nRF chips when detected. The supported gestures
include wrist-twist, double-tap and single-tap.

This commit implements optional raise-to-wake support based on such a
hardware feature. However, for compatibility with earlier units, which
has a BMA421 sensor that does not support gesture detection, we also
introduce a software-based detection algorithm for fallback. This
algorithm is based on the InfiniTime equivalent
<https://github.com/InfiniTimeOrg/InfiniTime/pull/826>.

Due to the way the accelerometer is mounted inside the PineTime, axes
remapping is needed for the hardware to detect the gestures correctly.
This was already implemented in the bma42x driver, and only needs to be
exposed to the Python side via bma42x-upy, which will have a
corresponding pull request for that.

Note that the hardware will never remap the raw data readout -- the
remapping information provided to the hardware will only be used in the
hardware-based algorithms, such as step counting and gesture detection.

Previously in #275, software-based axes remapping was already
introduced, however the mapping was inconsistent with what the hardware
expects. The hardware expects a left-handed coordinate system flipped
around the origin: when worn on the left wrist, the directions down, left,
and towards the user should be positive. On the PineTime, this translates
to all axes being flipped *and* the x and y axes swapped.

The software mapping is now corrected to agree with what the hardware
expects. This also means that apps making use of the accelerometer (for
now, just the Level app) will need to be changed, but that is beyond the
scope of this commit.

Co-authored-by: Ashley Eastwood <aeastw.git@fastmail.co.uk>
Co-authored-by: Finlay Davidson <finlay.davidson@coderclass.nl>