Adding a Four In A Row app
This is the classic Four In A Row game, also known as Connect4.  You play against the computer and you can adjust the difficulty level.

This is essentially the same code as last year (2022).

Signed-off-by: Tony Robinson <tony@tonyRobinson.com>

Adding a Four In A Row app

This is the classic Four In A Row game, also known as Connect4.  You play against the computer and you can adjust the difficulty level.

This is essentially the same code as last year (2022).

Signed-off-by: Tony Robinson <tony@tonyRobinson.com>

Rename four_in_a_row.py to apps/four_in_a_row.py

Fixing: The app and the screenshot should be moved in the appropriate subdirectories

Signed-off-by: Tony Robinson <tony@tonyRobinson.com>

fixing file location

Signed-off-by: Tony Robinson <tony@tonyRobinson.com>

Delete FourInARowApp.png

Signed-off-by: Tony Robinson tony@tonyRobinson.com
Signed-off-by: Tony Robinson <tony@tonyRobinson.com>

Forgot to update apps.rst

Signed-off-by: Tony Robinson <tony@tonyRobinson.com>

removed global wwdraw in favour of the standard local 'draw = wasp.watch.drawable'

Signed-off-by: Tony Robinson <tony@tonyRobinson.com>

Somehow I messed this file up

Signed-off-by: Tony Robinson <tony@tonyRobinson.com>